refactor: add common path for playground app entries
