+++ packages/playground/apps/_common/components/side-panel.ts
